backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.3.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
stripe==12.5.0
pytest==8.4.2
pytest-cov==6.3.0
httpx==0.28.1
